when converting microformats2, skip non-URL image properties
we've seen a couple of instances where photos are marked up with both p-photo
and u-photo, resulting in non-URL text in the photo which gives an error when
trying to publish. since we can count on mf2py to return absolute URLs for u-
properties, we can just skip any url that doesn't have a netloc.

pertains to https://github.com/snarfed/bridgy/issues/296