mf2 to html: de-dupe image and attachments
fixes #120...but by dropping the <img> *outside* content, not the one inside as requested. that may still be the right long term answer, but right now it's reused as atom content, and we're still debating whether we want to move that to an enclosure in #113 and snarfed/instagram-atom#3. reader support seems unclear at best.