Add RSVP library support for Meetup.com
As part of snarfed/bridgy/issues/873 we want to add Meetup.com support
to Brid.gy.

Before we can do this, we need to onboard it to granary.

For now, all were adding is RSVP support, so we can't add that in the
Granary webapp, just the library.

Unfortunately we don't get a permalink / ID back from the RSVP endpoint,
so similar to the way that Granary's Facebook likes work, we create our
own "permalink".