Tag RSVP with the Actor URL not `numeric_id`
Unfortunately it's not guaranteed for the actor's `numeric_id` to be
present, i.e. when Brid.gy publishes a post, as the actor will be a URL
i.e. https://www.jvt.me.

One suggestion from snarfed is to have Brid.gy pass through the ID from
the `MeetupAuth`, but a quicker win is that we can use the URL-encoded
profile URL of the user.