Fix: Escape HTML content before embedding in a post
As part of bridgy#880, we added a check to ensure that we would not be
subject to Cross-Site Scripting attacks (XSS).

However, as noted in bridgy#936, a post that does contain HTML is still
being rejected.

The simplest way to get around this is to escape the HTML tags, so they
can be safely rendered.

Test data via https://stackoverflow.com/a/5072031 adding a new class
which embeds our `content` property.

Closes bridgy#936.