Fix: Correctly trim Twitter alt text
As noted in [0], Twitter alt text is getting pre-emptively trimmed.

This is a fixup for the original introduction in 9a61adc5, which did not
pass in a `words`, therefore the alt text is getting trimmed at the
default 14 words.

We can use 1000 words, as it's very unlikely to be met before 420
characters in the alt text.

[0]: https://github.com/snarfed/bridgy/issues/1175