Correct the AO3 code, so it will work with bookmarks
The class of the works on the bookmarks page has the class of a bookmark
instead of a work. But there is still a class indicating the work ID there. So
I am changing the parsing to try to find the work ID in the classes, instead of
a work class.